backend/requirements.txt
flask
flask-cors
torch
transformers
pillow
